dcp-815 separate node_set.py
